+++ source/interface/IRenderable.cpp
@@ -101,14 +101,14 @@ INLINE BOOL IRenderable::IsVisible() const
 	return bVisible;
 }
 
-INLINE void IRenderable::SetColor(u8 r, u8 g, u8 b, u8 a)
+INLINE void IRenderable::SetColor(u32 r, u32 g, u32 b, u32 a)
 {
 	iColor = PIXEL_COLOR(r, g, b, a);
 }
 
 INLINE void IRenderable::SetColor(f32 r, f32 g, f32 b, f32 a)
 {
-	iColor = PIXEL_COLOR(static_cast<u8>(r * 255), static_cast<u8>(g * 255), static_cast<u8>(b * 255), static_cast<u8>(a * 255));
+	iColor = PIXEL_COLOR(static_cast<u32>(r * 255), static_cast<u32>(g * 255), static_cast<u32>(b * 255), static_cast<u32>(a * 255));
 }
 
 INLINE void IRenderable::SetColor(PIXEL px)
